LPS-160588 Reorganize directory and remove unecessary components
