--- modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-partner-portal/extra/remote-app/src/routes/MDFRequestForm/steps/Activities/Form/components/BudgetBreakdownSection/utils/getNewBudget.ts
@@ -1,20 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * The contents of this file are subject to the terms of the Liferay Enterprise
- * Subscription License ("License"). You may not use this file except in
- * compliance with the License. You can obtain a copy of the License by
- * contacting Liferay, Inc. See the License for the specific language governing
- * permissions and limitations under the License, including but not limited to
- * distribution rights of the Software.
- */
-
-import MDFRequestBudget from '../../../../../../../../common/interfaces/mdfRequestBudget';
-
-export default function getNewBudget(): MDFRequestBudget {
-	return {
-		cost: '',
-		expense: {},
-		r_activityToBudgets_c_activityId: '',
-	};
-}
