--- modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-partner-portal/extra/remote-app/src/routes/MDFRequestForm/steps/Activities/Form/utils/getNewActivity.ts
@@ -1,42 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * The contents of this file are subject to the terms of the Liferay Enterprise
- * Subscription License ("License"). You may not use this file except in
- * compliance with the License. You can obtain a copy of the License by
- * contacting Liferay, Inc. See the License for the specific language governing
- * permissions and limitations under the License, including but not limited to
- * distribution rights of the Software.
- */
-
-export default function getNewActivity() {
-	return {
-		activityPromotion: '',
-		ad: '',
-		assetsLiferayRequired: '',
-		budgets: [],
-		description: '',
-		endDate: '',
-		gatedLandingPage: '',
-		goalOfContent: '',
-		howLiferayBrandUsed: '',
-		keywordsForPPCCampaigns: '',
-		leadFollowUpStrategies: [],
-		leadGenerated: '',
-		liferayBranding: '',
-		liferayParticipationRequirements: '',
-		location: '',
-		marketingActivity: '',
-		mdfRequestAmount: '',
-		name: '',
-		overallMessageContentCTA: '',
-		primaryThemeOrMessage: '',
-		r_tacticToActivities_c_tacticId: '',
-		r_typeActivityToActivities_c_typeActivityId: '',
-		sourceAndSizeOfInviteeList: '',
-		specificSites: '',
-		startDate: '',
-		targetofLeads: '',
-		totalCostOfExpense: '',
-	};
-}
