+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-partner-portal/extra/remote-app/src/routes/MDFRequestForm/steps/Activities/components/Form/components/MiscellaneousMarketingFields/MiscellaneousMarketingFields.tsx
@@ -1,0 +1,27 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * The contents of this file are subject to the terms of the Liferay Enterprise
+ * Subscription License ("License"). You may not use this file except in
+ * compliance with the License. You can obtain a copy of the License by
+ * contacting Liferay, Inc. See the License for the specific language governing
+ * permissions and limitations under the License, including but not limited to
+ * distribution rights of the Software.
+ */
+
+import PRMForm from '../../../../../../../../common/components/PRMForm';
+import PRMFormik from '../../../../../../../../common/components/PRMFormik';
+
+interface IProps {
+	currentActivityIndex: number;
+}
+
+const MiscellaneousMarketingFields = ({currentActivityIndex}: IProps) => (
+	<PRMFormik.Field
+		component={PRMForm.InputText}
+		label="Describe the marketing activity"
+		name={`activities[${currentActivityIndex}].marketingActivity`}
+	/>
+);
+
+export default MiscellaneousMarketingFields;
